+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ## [next]
 
-- fix(): setupCurrentTransform/Control mousedown unexpected behavior [#9842](https://github.com/fabricjs/fabric.js/pull/9842)
+- dep(Canvas): `_setupCurrentTransform` => `setupCurrentTransform` fixing control mousedown unexpected behavior [#9842](https://github.com/fabricjs/fabric.js/pull/9842)
 - ci(): Add Jest coverage to the report [#9836](https://github.com/fabricjs/fabric.js/pull/9836)
 - test(): Add cursor animation testing and migrate some easy one to jest [#9829](https://github.com/fabricjs/fabric.js/pull/9829)
 - fix(Group, Controls): Fix interactive group actions when negative scaling is involved [#9811](https://github.com/fabricjs/fabric.js/pull/9811)