+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## [next]
 
+- test(): add event pointer caching failing test [#9850](https://github.com/fabricjs/fabric.js/pull/9850)
 - fix(Canvas): mouse move before event data [#9849](https://github.com/fabricjs/fabric.js/pull/9849)
 - chore(FabricObject): pass `e` to `shouldStartDragging` [#9843](https://github.com/fabricjs/fabric.js/pull/9843)
 - ci(): Add Jest coverage to the report [#9836](https://github.com/fabricjs/fabric.js/pull/9836)